Fix first round of compile errors
The build now reaches the compiler. Three failures:

ensure_resampler() moved into cloudvocal-processing.cpp but that file
never included cloudvocal-utils.h, where convert_speaker_layout lives.

obs_scene_sceneitem_from_source is deprecated in OBS 31 and CI builds
with warnings-as-errors. obs_scene_add already returns the scene item, so
the lookup was redundant; use its return value directly. Note that item
is scene-owned, so the matching release had to go - the deprecated call
returned a new reference, obs_scene_add does not.

HMAC takes key length as int; pass the size_t explicitly cast.

+++ src/utils/ssl-utils.cpp
@@ -42,7 +42,8 @@ std::string hmacSha256(const std::string &key, const std::string &data, bool isH
 		keyBytes.assign(key.begin(), key.end());
 	}
 
-	if (!HMAC(EVP_sha256(), keyBytes.data(), keyBytes.size(), (unsigned char *)data.c_str(),
+	// HMAC takes the key length as int; keyBytes is small (a hex or raw API key).
+	if (!HMAC(EVP_sha256(), keyBytes.data(), (int)keyBytes.size(), (unsigned char *)data.c_str(),
 		  data.length(), digest, (unsigned int *)&len)) {
 		obs_log(LOG_ERROR, "hmacSha256 failed during HMAC operation");
 		return {};
